requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
django-unfold==0.74.1
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
idna==3.11
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.6.2
